Added support for parameter overloading, consolidated prettyprinting into .ToString. Going to add error handling where it'll show you the line and code snippet of where a fault was found.

+++ Lexer/Objects/LexToken.cs
@@ -21,6 +21,6 @@ namespace CommonC.Lexer.Objects
         /// <summary>
         /// Line info
         /// </summary>
-        public int Line { get; set; }
+        public ulong Line { get; set; }
     }
 }
